requirements.txt
chainlit==1.3.0rc1
openai
yfinance
plotly
atlassian-python-api
pyodbc
websockets
azure-search-documents